ci: extract buildShapeMaterial + update relower review-fetch tests
CI red on PR #315:
- lint: ShapeGeometry.tsx exported buildShapeMaterial alongside the
  component, which violates react-refresh/only-export-components. Moved
  the helper into its own buildShapeMaterial.ts module and re-imported.
- test: 2 GeometryContext tests asserted the OLD skipReview-on-SSE-relower
  behavior that this PR deliberately dropped (so HUD interference count
  updates live on param drags). Renamed and updated assertions to expect
  the new (mesh, review) pair pattern: solo relower fires 2 fetches not 1;
  burst-of-3 collapses to (active mesh+review) + (trailing mesh+review).

13 tests pass after the fix.
